refactor: split validate from context, move to validator
